LPS-86148 Move lodash to it's own module
