+++ modules/apps/frontend-js/frontend-js-lodash-web/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Frontend JS Lodash Web
+Bundle-SymbolicName: com.liferay.frontend.js.lodash.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /frontend-js-lodash-web
+-includeresource: META-INF/resources=tmp/META-INF/resources
