+++ modules/apps/frontend-js/frontend-js-lodash-web/package.json
@@ -1,0 +1,7 @@
+{
+	"dependencies": {
+		"lodash": "4.17.4"
+	},
+	"name": "frontend-js-lodash-web",
+	"version": "1.0.0"
+}
