+++ modules/apps/frontend-js/frontend-js-lodash-web/src/main/resources/content/Language.properties
@@ -1,0 +1,3 @@
+enable-lodash=Enable lodash
+frontend-js-lodash-configuration-name=JavaScript Lodash
+frontend-js-lodash-description=For changes made here to take effect, you must reload the browser page after saving.
