+++ modules/apps/frontend-js/frontend-js-web/package.json
@@ -20,7 +20,6 @@
 		"clay-alert": "2.4.1",
 		"jquery": "^3.3.1",
 		"liferay-amd-loader": "3.1.2",
-		"lodash": "4.17.4",
 		"lodash.escape": "4.0.1",
 		"lodash.groupby": "4.6.0",
 		"lodash.isequal": "4.5.0",
